LayoutTests/platform/glib/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-referrer-policy.https-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://web-platform.test:9443/service-workers/service-worker/resources/simple.html?referrerPolicy was not allowed to display insecure content from http://web-platform.test:8800/service-workers/service-worker//resources/simple.html?referrerFull.

CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/service-workers/service-worker//resources/simple.html?referrerFull due to access control checks.


FAIL Service Worker responds to fetch event with the referrer policy promise_test: Unhandled rejection with value: object "TypeError: Load failed"
PASS Service Worker should respond to fetch with the default referrer policy
PASS Service Worker should respond to fetch with the referrer URL when a member of RequestInit is present - Default Referrer
